+++ src/parallax/tool/player/main.go
@@ -15,7 +15,7 @@ var optData = flag.String("data", "./data", "Directory with FCTP data files")
 var optPreload = flag.Bool("load", true, "Load all data files (instances)")
 var optThreads = flag.Int("threads", runtime.NumCPU(), "Number of system threads")
 var optEngine = flag.String("engine", engine.BID_GUROBI_EDGES, "Engine Name (RandomEdges, FirstEdges, ...)")
-var optFactor = flag.Int("factor", 2, "Price multiplication factor (Variable cost)")
+var optFactor = flag.Float64("factor", 2., "Price multiplication factor (Variable cost)")
 var verbose = flag.Int("verbose", 1, "Print a lot of messages, level 0, 1, 2, 3")
 
 func main() {
